1.7.8
* Added the Link Label On Path Dragging sample, demonstrating how to
restrict the dragging of link labels to remain on the path of the link.
This makes use of the LinkLabelOnPathDraggingTool.js extension.
* TableRows and TableColumns now update their ports correctly when
scrolling by setting Panel.topIndex or Panel.leftIndex.
* Links that have TwoWay Bindings of Link.points now copy faithfully
between different Diagrams.
* When the DraggingTool.dragsLink is true, a drag of a Link that
reconnects one or both ends of the Link will now raise one or two
"LinkRelinked" DiagramEvents. Just as with a "LinkRelinked" event raised
by the RelinkingTool, the DiagramEvent.parameter will be the old port
with which the Link had been connected.