1.7.16
• Translated and modularized the extensions into TypeScript *.ts files,
in the extensionsTS directory. The compiled files are present in that
directory as *.js files, which can be required or imported even if you
are not using TypeScript. (Caution: all of the script in the HTML
samples in the extensions directory have been split out into
...Script.ts files, so try not to be confused by the corresponding *.js
files, which will have Script in their names.)
• Fixed missing calls to Node.linkConnected and Node.linkDisconnected
event handlers when connecting or disconnecting reflexive links.
• Fixed error in Model.applyIncrementalJson adding group nodes.