1.7.22
* Improved routing of links to non-visible member nodes of collapsed
groups, when the spots of the group's default port are different than
thespots of the non-visible node that the link is connected with.
* Diagram mouseOut now appropriately calls Tool.standardMouseOver, and
Tool.standardMouseOver uses null as the current object if the mouse is
not over the current Diagram.
* For wrapping, TextBlocks respect the maxSize of their Panel, but also
accidentally conformed to their Panel's minSize.
* Diagram keyDown used to always preventDefault on keyDown events, but
this is now overridable via InputEvent.bubbling.