1.7.25
* Replaced the Angular 2 project with a newer Angular 4 project, at
../projects/angular-basic. This projected was created with Angular CLI;
we then added a diagram-editor Component consisting of a Diagram and a
Palette. That component takes a Model as an Input and produces
nodeSelected and modelChanged as Outputs. The main app component shows
editable details for the primary selection if it is a Node.
* Automatically redraw with improved sharpness as the user zooms the
page.
* Fixed a bug where Pictures given a GraphObject.maxSize and no loaded
Picture.element or Picture.source would measure incorrectly.