1.8.7
* Fixed routing of AvoidsNodes Links when added as members of a Group
that hadn't been measured fully yet with its new members.
* Fixed Layout.doLayout, the default layout that only positions nodes
that do not have a real Part.location, to respect the pre-layout
position of Groups, when the Group.layout has not been set.
* Improved LinkingBaseTool.isValidCycle to guess whether
LinkingTool.archetypeLinkData will create a new Link that is not
Link.isTreeLink, which affects link validity when Diagram.validCycle is
Diagram.CycleDestinationTree or Diagram.CycleSourceTree.
* Increased the CSS z-index of the default touch context menu from 300
to 10002.
* For Table Panels and RowColumnDefinitions: Before, a separator would
exist and take up space for every row and column except index = 0. Now
they exist and take up space for every row and column except the first
row and column, whatever its index may be.