change select default value from blank to text
Hi!
default select option as blank (causes an error),

if you choose a (blank) option or you didn't select an option for the first time, you will not be able to change or select another value. 
for that, I created a variable named (Val) to handle this error!

wonderful library!